+++ NEWS.md
@@ -1,4 +1,6 @@
 # srvyr 0.3.2.9000
+* Add warning to explain that design effects cannot be calculated on
+  proportions. (#39, thanks @mlaviolet)
 
 # srvyr 0.3.2
 * Bug fix for calculating multiple quantiles on grouped data